+++ OnlineBookClub.WEB/Models/DB/Const/Level.cs
@@ -6,7 +6,10 @@ namespace OnlineBookClub.WEB.Models.DB.Const
     {
         [Key]
         public byte LevelId { get; set; }
-        public int ExperiencePoint { get; set; }
+
+        public Int16 ExperiencePoint { get; set; }
+
+        //?=========> REFERANCES
 
         public virtual List<Achievement> Achievements  { get; set; }
